fix snapshot, fix signature

+++ crates/fuel-core/src/executor.rs
@@ -3272,7 +3272,7 @@ mod tests {
         }
 
         impl fuel_core_executor::ports::TransactionsSource for BadTransactionsSource {
-            fn next(&self, _: u64, _: u16, _: u32) -> Vec<MaybeCheckedTransaction> {
+            fn next(&self, _: u64, _: u16, _: u64) -> Vec<MaybeCheckedTransaction> {
                 std::mem::take(&mut *self.transactions.lock().unwrap())
             }
         }
